Made correct delete and update logic

+++ src/payments/payments.service.ts
@@ -37,7 +37,7 @@ export class PaymentsService {
       userId: userId
     })
 
-    await this.balanceService.updateBalance(userId, payment)
+    await this.updateBalance(payment)
 
     return payment;
   }
@@ -46,16 +46,106 @@ export class PaymentsService {
     return await this.paymentRepository.findAll(userId);
   }
 
+  private updateBalance(payment: Payment) {
+    let change;
+    if(payment.type === 'INCOME') change = payment.amount
+    else change = -payment.amount;
+
+    return this.balanceService.updateBalance(payment.userId, change);
+  }
+
   async getOneById(id: string) {
     return await this.paymentRepository.findUnique({ id })
   }
 
   async updatePayment(data: UpdatePaymentDTO, id: string) {
-    return await this.paymentRepository.updateOne(id, data)
+    const payment = await this.paymentRepository.findUnique({ id });
+    const balance = await this.balanceService.getBalance(payment.userId);
+  
+    const updatedBalance = this.calculateBalanceAfterUpdate(payment, data, balance.balance);
+    console.log(updatedBalance)
+    if (updatedBalance < 0) {
+      throw new BadRequestException('Updating this payment would result in a negative balance.');
+    }
+
+    const updatedPayment = await this.paymentRepository.updateOne(id, data);
+    this.updateBalanceOnUpdate(payment, data, balance.balance, updatedPayment);
+  
+    return updatedPayment;
   }
+  
+  private updateBalanceOnUpdate(payment: Payment, updatedData: UpdatePaymentDTO, balance: number, updatedPayment: Payment) {
+    let change = 0;
+    if (updatedData.amount !== undefined) {
+      change = updatedData.amount - payment.amount;
+    }
+  
+    if (updatedData.type !== undefined && updatedData.type !== payment.type) {
+      if (updatedData.type === 'EXPENSE' && payment.type === 'INCOME' && (balance - payment.amount) < 0) {
+        throw new BadRequestException('Changing payment type from income to expense would result in a negative balance.');
+      }
+      if (updatedData.type === 'INCOME' && payment.type === 'EXPENSE' && (balance + payment.amount) < updatedPayment.amount) {
+        throw new BadRequestException('Changing payment type from expense to income would not make balance positive.');
+      }
+    }
+  
+    return this.balanceService.updateBalance(payment.userId, change);
+  }
+  
+  private calculateBalanceAfterUpdate(payment: Payment, updatedData: UpdatePaymentDTO, currentBalance: number): number {
+    let updatedBalance = currentBalance;
+  
+    if (updatedData.type !== undefined && updatedData.type !== payment.type) {
+      if (updatedData.type === 'EXPENSE' && payment.type === 'INCOME') {
+        updatedBalance -= payment.amount *2;
+      } else if (updatedData.type === 'INCOME' && payment.type === 'EXPENSE') {
+        updatedBalance += payment.amount*2;
+      }
+    }
+  
+    if (updatedData.amount !== undefined && updatedData.amount !== payment.amount) {
+      const amountChange = updatedData.amount - payment.amount;
+      updatedBalance -= payment.type === 'EXPENSE' ? amountChange : -amountChange;
+    }
+  
+    return updatedBalance;
+  }  
 
   async deletePayment(id: string) {
+    const payment = await this.paymentRepository.findUnique({ id });
+    const payments = await this.paymentRepository.findInTime(payment.createdAt, payment.userId);
+    const balance = await this.balanceService.getBalance(payment.userId);
+  
+    const updatedBalance = this.calculateBalanceAfterDelete(payment, payments, balance.balance);
+    if (updatedBalance < 0) {
+      throw new BadRequestException('Deleting this payment would result in a negative balance.');
+    }
+  
+    this.updateBalanceOnDelete(payment);
+
     return await this.paymentRepository.deleteOne(id);
+    
+  }
+
+  private updateBalanceOnDelete(payment: Payment) {
+    let change;
+    if(payment.type === 'EXPENSE') change = payment.amount
+    else change = -payment.amount;
+
+    return this.balanceService.updateBalance(payment.userId, change);
+  }
+  
+  private calculateBalanceAfterDelete(payment: Payment, payments: Payment[], currentBalance: number): number {
+    const previousPaymentsTotal = payments.reduce((total, p) => total + (p.type === 'EXPENSE' ? p.amount : -p.amount), 0);
+    
+    let updatedBalance = currentBalance + payment.amount;
+    if (payment.type === 'EXPENSE') {
+      updatedBalance += previousPaymentsTotal;
+    } else {
+      updatedBalance -= previousPaymentsTotal;
+    }
+  
+    return updatedBalance;
   }
 
   async getPaymentsAfterTime(date: Date, userId: string) {
